Consolidated the IdTable and IdTableRow files.
