+++ src/engine/Comparators.h
@@ -6,7 +6,7 @@
 #include <utility>
 #include <vector>
 
-#include "./IdTable.h"
+#include "engine/idTable/IdTable.h"
 
 using std::pair;
 using std::vector;
